added

+++ backend/routers/posts.py
@@ -11,6 +11,7 @@ from backend.database import Base, engine, get_db
 from typing import List
 from contextlib import asynccontextmanager
 from fastapi.exception_handlers import http_exception_handler, request_validation_exception_handler
+from backend.auth import CurrentUser
 
 router = APIRouter()
 
@@ -23,18 +24,11 @@ async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
 
 # /api/posts
 @router.post('', response_model=PostResponse, status_code=status.HTTP_201_CREATED)
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(select(UserModel).where(UserModel.id == post.user_id))
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
+async def create_post(post: PostCreate, current_user : CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     new_post = PostModel(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id,
     )
     db.add(new_post)
     await db.commit()
@@ -44,7 +38,7 @@ async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_
 
 @router.get('/{post_id}', response_model=PostResponse)
 async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(select(PostModel).where(PostModel.id == post_id))
+    result = await db.execute(select(PostModel).options(selectinload(PostModel.author)).where(PostModel.id == post_id))
     post = result.scalars().first()
     if post:
         return post
@@ -52,7 +46,7 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
 
 
 @router.put('/{post_id}', response_model=PostResponse)
-async def update_post_full(post_id, post_data: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_full(post_id, post_data: PostCreate, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(PostModel)
                               .options(selectinload(PostModel))
                               .where(PostModel.id == post_id))
@@ -60,16 +54,13 @@ async def update_post_full(post_id, post_data: PostCreate, db: Annotated[AsyncSe
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
     
-    if post_data.user_id != post.user_id:
-        result = await db.execute(select(UserModel).where(UserModel.id == post_data.user_id))
-        user = result.scalars().first()
-        
-        if not user:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail = "You don't have permissions to edit this post")
+    
     # full update
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
+    post.user_id = current_user.id
     
     await db.commit()
     await db.refresh(post)
@@ -77,11 +68,14 @@ async def update_post_full(post_id, post_data: PostCreate, db: Annotated[AsyncSe
 
 
 @router.patch('/{post_id}', response_model=PostResponse)
-async def update_post(post_id, post_data: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post(post_id, post_data: PostCreate, current_user : CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(PostModel).where(PostModel.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+    
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code = status.HTTP_403_FORBIDDEN, detail = "You dont have permission to edit this post")
     
     update_data = post_data.model_dump(exclude_unset=True)
     # exclude_unset=True filters the fields user actually send instead of populating with default fields
@@ -95,11 +89,14 @@ async def update_post(post_id, post_data: PostCreate, db: Annotated[AsyncSession
     
     
 @router.delete('/{post_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id, db : Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id, current_user : CurrentUser, db : Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(PostModel).where(PostModel.id == post_id))
     post = result.scalars().first()
     
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+    
+    if post.user_id == current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You dont have permission to delete this post")
     await db.delete(post)
     await db.commit()
